backend/.aws-sam/auto-dependency-layer/HighlighterFunction/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
astroid==3.0.2
autopep8==2.0.4
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.7
distro==1.9.0
docstring-to-markdown==0.13
exceptiongroup==1.2.0
flake8==6.1.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
isort==5.13.2
jedi==0.19.1
Levenshtein==0.23.0
mccabe==0.7.0
msgpack==1.0.7
mypy-extensions==1.0.0
openai==1.7.2
packaging==23.2
parso==0.8.3
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.3.0
pyflakes==3.1.0
pylint==3.0.3
pynvim==0.5.0
python-dotenv==1.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
pytoolconfig==1.3.1
rapidfuzz==3.6.1
requests==2.31.0
rope==1.11.0
sniffio==1.3.0
snowballstemmer==2.2.0
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
whatthepatch==1.0.5
yapf==0.40.2
zipp==3.17.0
